refactor(telegram): split commands.ts into command-group modules
commands.ts was 904 lines — one registerCommands function wiring ~16
bot.command() handlers that shared an ADMIN_USER_ID module global. Split
into a commands/ folder by category:
  - definitions — the TELEGRAM_COMMANDS menu
  - state       — admin-id holder + isAuthorizedAdmin guard
  - info        — /start /help /ping /memory /plugins
  - session     — /reset /status
  - settings    — /model /effort /pulse /settings
  - admin       — /admin /metrics /doctor /dream /soul /restart /update
                  + unknown-command suggester (registered last, unchanged)

index.ts re-exports TELEGRAM_COMMANDS + setAdminUserId and rebuilds
registerCommands wiring each group in the original order. Importer
updated to the barrel path.
